Refactor modules * Combine ui_mqtt_bridge and mqtt * Move AbstractStabilizerInterface and WavemeterInterface to interface.py * Make  directory a single file
