+++ stabilizer_ui/mqtt.py
@@ -1,25 +1,19 @@
 import asyncio
-from contextlib import suppress
-from typing import Any, Optional, Iterable
 import logging
 import json
-
-from gmqtt import Client as MqttClient
 import uuid
-import stabilizer
-import numpy as np
+
+from typing import NamedTuple, List, Callable, Any, Dict
+from contextlib import suppress
+from PyQt5 import QtWidgets
+from gmqtt import Client as MqttClient
+from .widgets import AbstractUiWindow
 
 logger = logging.getLogger(__name__)
-
-Y_MAX = stabilizer.voltage_to_machine_units(stabilizer.DAC_FULL_SCALE)
 
 
 def _int_to_bytes(i):
     return i.to_bytes(i.bit_length() // 8 + 1, byteorder="little")
-
-
-def starts_with(string, prefix) -> bool:
-    return len(string) >= len(prefix) and string[:len(prefix)] == prefix
 
 
 class MqttInterface:
@@ -125,69 +119,176 @@ class MqttInterface:
                 err.__cause__ = e
                 result.set_exception(err)
         return 0
+    
+
+class NetworkAddress(NamedTuple):
+    ip: List[int]
+    port: int = 9293
+
+    @classmethod
+    def unspecified(cls):
+        return cls.from_str_ip("0.0.0.0", 0)
+
+    @classmethod
+    def from_str_ip(cls, ip: str, port: int):
+        _ip = list(map(int, ip.split(".")))
+        return cls(_ip, port)
+
+    def get_ip(self) -> str:
+        return ".".join(map(str, self.ip))
 
 
-class AbstractStabilizerInterface:
-    """
-    Shim for controlling stabilizer over MQTT
-    """
+def read(widgets):
+    assert len(widgets) == 1, "Default read() only implemented for one widget"
+    widget = widgets[0]
 
-    def __init__(self):
-        self._interface_set = asyncio.Event()
-        self._interface: Optional[MqttInterface] = None
-
-    def set_interface(self, interface: MqttInterface) -> None:
-        self._interface = interface
-        self._interface_set.set()
-
-    async def change(self, *args, **kwargs):
-        await self._interface_set.wait()
-        await self.triage_setting_change(*args, **kwargs)
-
-    async def triage_setting_change(self):
-        raise NotImplementedError
-
-    async def set_pi_gains(self, channel: int, iir_idx: int, p_gain: float,
-                           i_gain: float):
-        b0 = i_gain * 2 * np.pi * stabilizer.SAMPLE_PERIOD + p_gain
-        b1 = -p_gain
-        await self.set_iir(channel, iir_idx, [b0, b1, 0, 1, 0])
-
-    async def set_iir(
-        self,
-        channel: int,
-        iir_idx: int,
-        ba: Iterable,
-        x_offset: float = 0.0,
-        y_offset: float = 0.0,
-        y_min: float = -Y_MAX,
-        y_max: float = Y_MAX,
+    if isinstance(
+            widget,
+        (
+            QtWidgets.QCheckBox,
+            QtWidgets.QRadioButton,
+            QtWidgets.QGroupBox,
+        ),
     ):
-        forward_gain = sum(ba[:3])
-        if forward_gain == 0 and x_offset != 0:
-            logger.warning("Filter has no DC gain but x_offset is non-zero")
-        key = f"{self.iir_ch_topic_base}/{channel}/{iir_idx}"
-        value = {
-            "ba": list(ba),
-            "u": stabilizer.voltage_to_machine_units(y_offset + forward_gain * x_offset),
-            "min": stabilizer.voltage_to_machine_units(y_min),
-            "max": stabilizer.voltage_to_machine_units(y_max),
-        }
-        await self.request_settings_change(key, value)
+        return widget.isChecked()
 
-    def publish_ui_change(self, topic: str, argument: Any):
-        payload = json.dumps(argument).encode("utf-8")
-        self._interface._client.publish(f"{self._interface._topic_base}/{topic}",
-                                        payload,
-                                        qos=0,
-                                        retain=True)
+    if isinstance(widget, (QtWidgets.QDoubleSpinBox, QtWidgets.QSpinBox)):
+        return widget.value()
 
-    async def request_settings_change(self, key: str, value: Any):
-        """
-        Write to the miniconf-provided topics, which currently returns a
-        string message as a reply; should really be JSON/… instead, see
-        quartiq/miniconf#32.
-        """
-        msg = await self._interface.request(key, value, retain=True)
-        if starts_with(msg, "Settings fail"):
-            logger.warning("Stabilizer reported failure to write setting: '%s'", msg)
+    if isinstance(widget, QtWidgets.QComboBox):
+        return widget.currentText()
+
+    assert f"Widget type not handled: {widget}"
+
+
+def write(widgets, value):
+    assert len(widgets) == 1, "Default write() only implemented for one widget"
+    widget = widgets[0]
+
+    if isinstance(
+            widget,
+        (
+            QtWidgets.QCheckBox,
+            QtWidgets.QRadioButton,
+            QtWidgets.QGroupBox,
+        ),
+    ):
+        widget.setChecked(value)
+    elif isinstance(widget, QtWidgets.QDoubleSpinBox):
+        widget.setValue(value)
+    elif isinstance(widget, QtWidgets.QComboBox):
+        options = [widget.itemText(i) for i in range(widget.count())]
+        widget.setCurrentIndex(options.index(value))
+    else:
+        assert f"Widget type not handled: {widget}"
+
+
+class UiMqttConfig(NamedTuple):
+    widgets: List[QtWidgets.QWidget]
+    read_handler: Callable = read
+    write_handler: Callable = write
+
+
+class UiMqttBridge:
+
+    def __init__(self, client: MqttClient, configs: Dict[Any, UiMqttConfig]):
+        self.client = client
+        self.configs = configs
+        self.panicked = False
+
+    @classmethod
+    async def new(cls, broker_address: NetworkAddress, *args, **kwargs):
+        will_message = kwargs.pop("will_message", None)
+        client = MqttClient(client_id="", will_message=will_message)
+        host, port = broker_address.get_ip(), broker_address.port
+        try:
+            await client.connect(host, port=port, keepalive=10)
+            logger.info(f"Connected to MQTT broker at {host}:{port}.")
+        except Exception as connect_exception:
+            logger.error("Failed to connect to MQTT broker: %s", connect_exception)
+            raise connect_exception
+
+        return cls(client, *args, **kwargs)
+
+    async def load_ui(self, objectify: Callable, root_topic: str, ui: AbstractUiWindow):
+        """Load current settings from MQTT"""
+        retained_settings = {}
+
+        def panic_handler(value):
+            has_panicked = (json.loads(value) is not None)
+            ui.onPanicStatusChange(has_panicked, value)
+            if has_panicked:
+                logger.error("Stabilizer had panicked, but has restarted")
+
+        def alive_handler(value, is_initial_subscription=False):
+            is_alive = bool(json.loads(value))
+            ui.onlineStatusChange(is_alive)
+            logger.info(f"Stabilizer {'alive' if is_alive else 'offline'}")
+
+        def collect_settings(_client, topic, value, _qos, _properties):
+            subtopic = topic[len(root_topic) + 1:]
+            try:
+                key = objectify(subtopic)
+                decoded_value = json.loads(value)
+                retained_settings[key] = decoded_value
+                logger.info(
+                    "Registering message topic '#/%s' with value '%s'",
+                    subtopic,
+                    decoded_value,
+                )
+                if subtopic == "meta/panic":
+                    panic_handler(value)
+                elif subtopic == "alive":
+                    alive_handler(value)
+            except ValueError:
+                logger.info("Ignoring message topic '%s'", subtopic)
+            return 0
+
+        self.client.on_message = collect_settings
+
+        logger.info(f"Subscribing to all settings at {root_topic}/#")
+        all_settings = f"{root_topic}/#"
+        self.client.subscribe(all_settings)
+        # Based on testing, all the retained messages are sent immediately after
+        # subscribing, but add some delay in case this is actually a race condition.
+        await asyncio.sleep(1)
+        self.client.unsubscribe(all_settings)
+
+        self.client.subscribe(f"{root_topic}/meta/panic")
+        self.client.subscribe(f"{root_topic}/alive")
+
+        for retained_key, retained_value in retained_settings.items():
+            if retained_key in self.configs:
+                cfg = self.configs[retained_key]
+                cfg.write_handler(cfg.widgets, retained_value)
+
+    def connect_ui(self):
+        """Set up UI signals"""
+        keys_to_write = set()
+        ui_updated = asyncio.Event()
+
+        # Capture loop variable.
+        def make_queue(key):
+
+            def queue(*args):
+                keys_to_write.add(key)
+                ui_updated.set()
+
+            return queue
+
+        for key, cfg in self.configs.items():
+            queue = make_queue(key)
+            for widget in cfg.widgets:
+                if not widget:
+                    continue
+                elif hasattr(widget, "valueChanged"):
+                    widget.valueChanged.connect(queue)
+                elif hasattr(widget, "toggled"):
+                    widget.toggled.connect(queue)
+                elif hasattr(widget, "activated"):
+                    widget.activated.connect(queue)
+                else:
+                    assert f"Widget type not handled: {widget}"
+
+            keys_to_write.add(key)  # write once at startup
+        return keys_to_write, ui_updated
